LayoutTests/platform/chromium/http/tests/security/no-popup-from-sandbox-top-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate a navigation change for frame with URL http://127.0.0.1:8000/security/no-popup-from-sandbox-top.html from frame with URL data:text/html,       <script>       var win = window.open('about:blank', '_top');       alert(win ? 'FAIL' : 'PASS');       </script>.

ALERT: PASS
To run this test outside of DumpRenderTree, please disable your popup blocker!
